add ui for answer item
