+++ lib/presentation/ui_config/global_widgets/user_avatar.dart
